Initial refactor and restyle, panel need to be styled
