+++ src/components/List/list.scss
@@ -1,0 +1,28 @@
+.list{
+  margin: 2%;
+  h1{
+    letter-spacing: 0.7px;
+    font-weight : 700;
+  }
+  h3{
+    margin-top: -10px;
+    letter-spacing: 0.7px;
+    font-weight : 500;
+    text-align: center;
+  }
+  img{
+    margin-top:15px;
+    max-width: 100%;
+    display: block;
+  }
+  .not-found{
+    img{
+      height: 300px;
+    }
+  }
+  .homepage{
+    img{
+      height: 250px;
+    }
+  }
+}
