+++ src/components/ResultRow/result-row.scss
@@ -1,0 +1,46 @@
+.result-row{
+    background-color: #fff;
+    line-height: 1.5rem;
+    padding: 15px 5px;
+    margin: 0;
+    margin-top:20px;
+    border-bottom: 1px solid #e0e0e0;
+
+    &:last-child{
+      border-bottom: 0px;
+    }
+    h3{
+      font-size: 22px;
+      letter-spacing: 0.7px;
+      font-weight: 700;
+      padding-bottom: 2px;
+      border-bottom: 2px dotted #B6B6B6;
+    }
+    p{
+      letter-spacing: 0.7px;
+      margin-top: -10px;
+      font-size: 14px;
+      color: #727272;
+    }
+    button {
+      display: inline-block;
+      position: relative;
+      width: 120px;
+      height: 32px;
+      right: 42px;
+      padding-bottom: 5px;
+      line-height: 27px;
+      border: 1px #455A64;
+      border-radius: 2px;
+      font-size: 0.9em;
+      background-color: #607D8B;
+      color: #fff;
+      transition: box-shadow 0.3s cubic-bezier(0.4, 0, 0.2, 1);
+      transition-delay: 0.3s;
+      box-shadow: 0 2px 5px 0 rgba(0, 0, 0, 0.26);
+    }
+    button:hover {
+      box-shadow: 0 6px 14px 0 rgba(0, 0, 0, 0.2);
+      transition-delay: 0s;
+    }
+}
