+++ src/components/SearchBar/search-bar.scss
@@ -1,4 +1,4 @@
-$color: #63C3E8;
+$color: #03A9F4;
 
 .searchBar{
   input::-webkit-input-placeholder, button {
@@ -8,9 +8,10 @@ $color: #63C3E8;
   input {
     margin: 40px 10px;
     width: 300px;
+    letter-spacing: 0.7px;
     display: inline-block;
     border: none;
-    padding: 10px 0;
+    padding: 12px 0;
     border-bottom: solid 1px $color;
     transition: all 0.3s cubic-bezier(.64,.09,.08,1);
     background: linear-gradient(to bottom, rgba(255,255,255,0) 96%, $color 4%);
@@ -23,7 +24,7 @@ $color: #63C3E8;
       outline: none;
       background-position: 0 0;
       &::-webkit-input-placeholder {
-        color: $color;
+        color: #fff;
         font-size: 12px;
         transform: translateY(-20px);
         visibility: visible !important;
@@ -35,12 +36,12 @@ $color: #63C3E8;
     background: $color;
     cursor: pointer;
     padding: 10px;
-    color: #ddd;
+    color: #fff;
     border-radius: 100%;
     margin-left: 0px;
     box-shadow: 0 3px 6px 0 rgba(0,0,0,0.2);
     &:hover {
-      transform: translateY(-1px);
+      transform: translateY(-3px);
       box-shadow: 0 5px 5px 0 rgba(0,0,0,0.2);
     }
   }
